test(trie): pin writer-reader agreement across the trie pipeline
The dict/BPE string encoder (writer) and decodeTrieDict (reader) are
hand-synced implementations of one format, and the streaming trie
descent was the only reader without a full-map test. Pin all of it:

- Extract the base-91/BPE encoder from write-decode-map.ts into
  scripts/trie/encode-dict.ts (unchanged logic; the script wrote files
  at import time, so the encoder wasn't importable from a spec).
  Generated files are byte-identical.
- New encode-dict.spec.ts round-trips the writer's encoding through the
  runtime decodeTrieDict: both real map tries, BPE-heavy patterns,
  delta escape/double-escape and RLE-chunking edge values, and the
  whole dictSize search grid.
- New exhaustive streaming spec: every entity from entities.json and
  xml.json through EntityDecoder, whole and char-by-char, asserting
  output and consumed match decodeHTML/decodeXML.
- Staleness guard in write-decode-map.ts: BPE_RANK_OVERRIDES were tuned
  against the exact current trie contents, so assert the encoded HTML
  trie length against a recorded constant and tell future editors to
  re-tune when it fires.
