+++ src/decode-stream.spec.ts
@@ -1,7 +1,33 @@
 import { describe, expect, it, vi } from "vitest";
-import { DecodingMode, EntityDecoder } from "./decode.js";
+import entityMap from "../maps/entities.json" with { type: "json" };
+import xmlMap from "../maps/xml.json" with { type: "json" };
+import { decodeHTML, decodeXML, DecodingMode, EntityDecoder } from "./decode.js";
 import { htmlDecodeTree } from "./generated/decode-data-html.js";
 import { xmlDecodeTree } from "./generated/decode-data-xml.js";
+
+/**
+ * Decode `&name;` through EntityDecoder in `chunkSize`-char writes.
+ * @param decodeTree The trie to decode against.
+ * @param input Full entity text, starting at the `&`.
+ * @param chunkSize Characters per write call.
+ */
+function streamEntity(
+    decodeTree: Uint16Array,
+    input: string,
+    chunkSize: number,
+): { output: string; consumed: number } {
+    let output = "";
+    const decoder = new EntityDecoder(decodeTree, (cp) => {
+        output += String.fromCodePoint(cp);
+    });
+    decoder.startEntity(DecodingMode.Legacy);
+    let consumed = -1;
+    for (let pos = 1; pos < input.length && consumed < 0; pos += chunkSize) {
+        consumed = decoder.write(input.slice(pos, pos + chunkSize), 0);
+    }
+    if (consumed < 0) consumed = decoder.end();
+    return { output, consumed };
+}
 
 describe("EntityDecoder Streaming", () => {
     it("should decode long entities split across chunks (char-by-char)", () => {
@@ -211,6 +237,37 @@ describe("EntityDecoder Streaming", () => {
             decoder.startEntity(DecodingMode.Attribute);
             expect(decoder.write(`${entity}=`, 1)).toBe(0);
             expect(callback).not.toHaveBeenCalled();
+        });
+    });
+
+    /*
+     * Exhaustive writer↔reader agreement for the streaming descent: it is
+     * the only trie reader without a full-map test otherwise. Every entity
+     * from both maps goes through EntityDecoder whole and char-by-char, and
+     * must match the sync decoders in both output and consumed count.
+     */
+    describe("exhaustive full-map agreement with the sync decoders", () => {
+        const chunkSizes = [
+            ["whole", Number.MAX_SAFE_INTEGER],
+            ["char-by-char", 1],
+        ] as const;
+
+        it.each(chunkSizes)("should decode every HTML entity (%s)", (_name, chunkSize) => {
+            for (const name of Object.keys(entityMap)) {
+                const input = `&${name};`;
+                const result = streamEntity(htmlDecodeTree, input, chunkSize);
+                expect(result.output).toBe(decodeHTML(input));
+                expect(result.consumed).toBe(input.length);
+            }
+        });
+
+        it.each(chunkSizes)("should decode every XML entity (%s)", (_name, chunkSize) => {
+            for (const name of Object.keys(xmlMap)) {
+                const input = `&${name};`;
+                const result = streamEntity(xmlDecodeTree, input, chunkSize);
+                expect(result.output).toBe(decodeXML(input));
+                expect(result.consumed).toBe(input.length);
+            }
         });
     });
